requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
distro==1.9.0
fastapi==0.115.12
Flask==3.1.1
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ollama==0.4.8
openai==1.78.0
playwright==1.52.0
pydantic==2.11.4
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.1
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
Werkzeug==3.1.3
